safety/efficiency update
removed all usage of new/delete - changed to std::unique_ptr for memory leak safety.

merged the 2 dynamic allocations in parallel crypto class into one.
this uses a new type that contains th has_data flag and the thread handle.

atomic operations were dropped from (default) seq_cst to acq/rel.
this is a boon for the optimizer but on my (windows) system there doesn't appear to be much of a difference.