changed from atomics to condition vars
the old impl used atomics to query when there was work to be done.
this has the problem of those threads not sleeping and thus using a lot of cpu.
this meant creating a parallel crypto object with a long lifespan was a SUPER bad idea in terms of performance.

new impl used condition vars for synchronization.
this has slightly lower performance due to additional mutexing and sleep stuff, but makes up for it in being very easy on the cpu when idle.
on my 4-core maching the previous version used ~80% cpu when idle, this one uses 0%.

additionally, with one of the new features that was added (moreso a semantic for the object's usage) i believe i can further increase performance with this model.