cleanup + reorganization
cleaned up the parallel crypto class a bit.
among other things, renamed it to crypto_service.

the file processing functions no longer take a crypto service argument - they're not member functions.

changed the key system to its own type.
this was to facilitate copy semantics for the keys, though this is currently unused.