type changes / typedefs
changed a bunch of int parameters (carried over from the C# impl) to std::size_t to be more C++y.

added a (private) typedef to the crypto_service class that controls what type to use for masks.
my intention with this is to use a faster type.
it's currently defined as int to be equivalent with what the last commit had.
this'll likely be changed later to std::fast_int8_t or something of the like - we only need 8 bits, aside from that it just needs to be fast.
there's also the question of cache line locality, so a smaller size would likely be better.